requirements.txt
google-ai-generativelanguage==0.6.4
langchain==0.2.1
langchain-google-genai==1.0.5
langchain-text-splitters==0.2.0
langchain-community==0.2.1
langchain-core==0.2.3
gradio==4.32
faiss-cpu==1.8